change the int to uint8_t of alignment encode for save space

+++ alignment/Alignment.h
@@ -44,7 +44,7 @@ public:
      * @param pattern_str Encoded sequence at a column (e.g., "0123")
      * @param freq How many times this pattern appears
      */
-    void addPattern(const std::vector<int> pattern_vec, int freq = 1) {
+    void addPattern(const std::vector<uint8_t> pattern_vec, int freq = 1) {
         patterns.emplace_back(pattern_vec, freq);
     }
